extracted dateToSeconds properly
